requirements.txt
asarPy==1.0.1
paho-mqtt==2.1.0
